refactor(s3mock): move from go/tools to go/test
s3mock is a mock S3 server used solely by the multipooler end-to-end backup
tests — its only importers are under go/test/. It doesn't fit go/tools/, which
is for generic, non-multigres-specific helpers. Relocating it to go/test/s3mock
reflects its actual role and brings it under the test-code gosec exclusions, so
its taint-analysis false positives (G706) no longer need annotating.
