+++ go/test/endtoend/multipooler/backupfaults/backup_archive_lag_test.go
@@ -29,8 +29,8 @@ import (
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 	multipoolermanagerdata "github.com/multigres/multigres/go/pb/multipoolermanagerdata"
 	"github.com/multigres/multigres/go/test/endtoend/shardsetup"
+	"github.com/multigres/multigres/go/test/s3mock"
 	"github.com/multigres/multigres/go/test/utils"
-	"github.com/multigres/multigres/go/tools/s3mock"
 )
 
 // TestPrimaryCrashWithUnarchivedWAL_NoDataLoss proves Multigres' synchronous-commit
