+++ go/test/endtoend/multipooler/backupfaults/backup_failover_test.go
@@ -26,8 +26,8 @@ import (
 
 	multipoolermanagerdata "github.com/multigres/multigres/go/pb/multipoolermanagerdata"
 	"github.com/multigres/multigres/go/test/endtoend/shardsetup"
+	"github.com/multigres/multigres/go/test/s3mock"
 	"github.com/multigres/multigres/go/test/utils"
-	"github.com/multigres/multigres/go/tools/s3mock"
 )
 
 // TestBackup_FailsDuringPrimaryFailover verifies that pgBackRest fails cleanly when
