+++ go/test/s3mock/server.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// go/tools/s3mock/server.go
+// go/test/s3mock/server.go
 
 // Package s3mock provides an S3-compatible mock server for testing.
 //
