+++ go/test/utils/process.go
@@ -21,6 +21,7 @@ import (
 	"time"
 
 	"github.com/multigres/multigres/go/common/constants"
+	"github.com/multigres/multigres/go/tools/ctxutil"
 	"github.com/multigres/multigres/go/tools/executil"
 )
 
@@ -75,7 +76,7 @@ func BaseTestEnv() []string {
 // on the returned Cmd before calling Start().
 func CommandWithOrphanProtection(monitorCtx context.Context, name string, args ...string) *executil.Cmd {
 	allArgs := append([]string{name}, args...)
-	cmd := executil.Command(context.Background(), "run_in_test.sh", allArgs...)
+	cmd := executil.Command(ctxutil.Detach(monitorCtx), "run_in_test.sh", allArgs...)
 	go func() {
 		<-monitorCtx.Done()
 		stopCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
